ac-atvx-personavix-main/api/requirements.txt
sqlalchemy
pydantic
fastapi
uvicorn
mysql-connector
mysqlclient
pylint
pytest
pylint_pytest
pylint_absolute_imports
httpx
mysqlclient
pytest-asyncio
python-jose==3.3.0
requests==2.32.2
prometheus_fastapi_instrumentator
python-dotenv
bcrypt